REFACTORED: Eureka microservice, had problems with spring cloud
