+++ microservice-eureka/pom.xml
@@ -11,7 +11,7 @@
 	<artifactId>microservice-eureka</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>microservice-eureka</name>
-	<description>Demo project for Spring Boot</description>
+	<description>Eureka microservice</description>
 	<url/>
 	<licenses>
 		<license/>
@@ -27,7 +27,7 @@
 	</scm>
 	<properties>
 		<java.version>21</java.version>
-		<spring-cloud.version>2024.0.1</spring-cloud.version>
+		<spring-cloud.version>2025.0.0</spring-cloud.version>
 	</properties>
 	<dependencies>
 		<dependency>
