+++ microservice-eureka/src/main/resources/banner.txt
@@ -1,0 +1,8 @@
+
+ ___ ___  ____   __  ____   ___   _____   ___  ____  __ __  ____   __    ___          ___  __ __  ____     ___  __  _   ____          ___   ____
+|   |   ||    | /  ]|    \ /   \ / ___/  /  _]|    \|  |  ||    | /  ]  /  _]        /  _]|  |  ||    \   /  _]|  |/ ] /    |        /   \ |    \
+| _   _ | |  | /  / |  D  )     (   \_  /  [_ |  D  )  |  | |  | /  /  /  [_ _____  /  [_ |  |  ||  D  ) /  [_ |  ' / |  o  | _____ |     ||  _  |
+|  \_/  | |  |/  /  |    /|  O  |\__  ||    _]|    /|  |  | |  |/  /  |    _]     ||    _]|  |  ||    / |    _]|    \ |     ||     ||  O  ||  |  |
+|   |   | |  /   \_ |    \|     |/  \ ||   [_ |    \|  :  | |  /   \_ |   [_|_____||   [_ |  :  ||    \ |   [_ |     \|  _  ||_____||     ||  |  |
+|   |   | |  \     ||  .  \     |\    ||     ||  .  \\   /  |  \     ||     |      |     ||     ||  .  \|     ||  .  ||  |  |       |     ||  |  |
+|___|___||____\____||__|\_|\___/  \___||_____||__|\_| \_/  |____\____||_____|      |_____| \__,_||__|\_||_____||__|\_||__|__|        \___/ |__|__|
