+++ microservice-eureka/src/test/java/com/microservice/eureka/microservice_eureka/MicroserviceEurekaApplicationTests.java
@@ -1,4 +1,4 @@
-package com.microservice.eureka;
+package com.microservice.eureka.microservice_eureka;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
